Move Index Update for Nav Cycle Methods to HarpoonList:Select
I changed the HarpoonList:Select() method to update the
HarpoonList:_index attribute to the selected index for consistency with
the behavior of harpoon1 when selecting an item from the quick menu.

The current behavior only updates the index when calling next() or
prev(), which creates some inconsistencies depending on the workflow.

This change causes consistent behavior when selecting an item from the
HarpoonList regardless of how/where it is selected.